+++ dataclass_args/__init__.py
@@ -113,7 +113,7 @@ from .exceptions import ConfigBuilderError, ConfigurationError, FileLoadingError
 from .file_loading import is_file_loadable_value, load_file_content
 from .utils import load_structured_file
 
-__version__ = "1.5.0"
+__version__ = "1.5.1"
 
 __all__ = [
     # Main API